+++ src/frontend/Navigation/Drawer.tsx
@@ -1,0 +1,17 @@
+import * as React from 'react';
+import {createDrawerNavigator} from '@react-navigation/drawer';
+import {RootStackNavigator} from './Stack';
+
+type DrawerScreens = {
+  DrawerHome: undefined;
+};
+
+export const Drawer = createDrawerNavigator<DrawerScreens>();
+
+export const DrawerNavigator = () => {
+  return (
+    <Drawer.Navigator>
+      <Drawer.Screen name="DrawerHome" component={RootStackNavigator} />
+    </Drawer.Navigator>
+  );
+};
